Add more port functions and support dynamic ports.
Previously the devices themselves used to created and
take care of ports. They only provided callback to get a
port based on integer.

Now devices only provide a callback to create a new port and
most of the port bookkeeping is done by the network. Support
for port classes and names was also added, so that the lsdn-load
can use them.