Add support for loading ports from YAML config.
Ports and porgroups are now loaded from file. There is not yet a way to
specify port attributes (such as fwd. table) and port connections are
ignored. Netdev device is also support so that we are able to load the
included example.

The YAML format is as we have discussed in MFF Lab, with following
changes:
1) Port groups are named group* instead of group.*
2) The name of the network is not derived from the key in the network
map, but instead is an optional name: key. In future i plan to derive
the name from the file name if not given.

The necessary support for YAML lists in config.{c,h} was added, port
groups in netmodel were implemented (or at least partially).

Also includes minor cleanup: char constization (please use next time),
using of abort (better debugging), bool options.

Hladová mlha honí a žere mladé oběti.