Static vxlan tunnels don't use the fdb table anymore.
Currently linux has reserved the special VNI value of 0
for the purposes of static vxlan tunnels where the tunnel
metadata are filled on-the-fly. (I didn't find any mention
about this behavior in any documentation, it was only
observed in the kernel source code, but I believe this
is the case. See e.g. the commit b9167b2e775 in the linux
kernel.) As there can be at most one vxlan vtep with the
same VNI connected to a single device that means we need
to reuse the same tunnel device for each static vxlan
tunnel hosted on the same machine.