Change the way sbridge works and add more netops.
These changes are in preparation for migration support. Sbridge (static
bridge) was changed from simple "commit sbridge" type calls to its own
module which keeps tracks of the participating interfaces, routes
associated with the interfaces and MAC address on the routes. Routes,
interfaces and MACs can be both added and removed, which is needed for
migrations.

The `commit_pa` netop was broken down to sub-ops (commit virt,
remote_pa, remote_virt) and corresponding remove ops were also added.
The static vxlan makes use of these ops. The common netmodel also now
contains more logic to track lifetime of the netmodel objects and also
keeps track of `remote` objects (virt and PAs). Thanks to this, there is
a simple mapping between the new objects/ops and sbridge ops for
vxlan_static:

 netmodel           | sbridge
 ---------------------------------------
 create_pa          = sbridge, sbridge interface for tunnel
 create_virt        = sbridge interface, route and MAC for the virt
 create_remote_pa   = sbridge route on the tunnel interface
 create_remote_virt = MAC on the tunnel route