Added support for user hooks.
Added support for user hooks with an example
of how one could use them to setup VMs with
a dhcp server and dhcp clients on a VLAN network.

If you would like to run the test yourself,
add in the / directory of the rootfs we use
for tests the domain-init script.

After the test network is setup, you can ping
from within a guest other VMs in the same network.
It would be nice to automate this task. What is also
missing at this point is the cleanup of all the
network interfaces.

The bulk of this commit is actually in the infrastructure.
The user hooks are to be improved as their
intended usage will become more evident.