Migrations.
Adds support for migrations in the vxlan_static network + corresponding
tests. Because of current limitations, the test just really commits the
network in different state and then tests the result (does not test the
state before migration).

This involved fixing various bugs in rules removal and actually calling
the remove_remove_virt and remove_virt operation. Support for remove_pa
and more types of network changes is not ready yet.