Improve cleanup of both memory and linux interfaces.
The commit phase was extended to include cleanup of all object. Also a
`decommit_disable` mode was added, which allows us to free the memory
structures without removing the interfaces. This mode is usefull for
testing memory leaks and also will be usefull once we get commandline
migrations without a daemon ready. Also a bunch of "_free" methods were
added.

The system is not yet perfect, was not valgrinded and we should also
think what should happen if we change certain attributes, like an IP
address of the phys. Do we want to disallow that completely, or mark the
whole phys for renew?

The tests now also deallocate memory after them (using
lsdn_context_free) and also there is a `cleanup` test part that calls
lsdn_context_cleanup and check that no helper lsdn- interfaces are left
over.

Also make note of TEST_RUNNER which you can set to "gdb --args" or
"valgrind" to run lsctl under those tools.