Workaround for vlan dhcp test.
With the transition to the new testing environment, where rootfs resides on 9p filesystem we are
facing a bug when you run this on top of COW FS, e.g. when you use btrfs as your main FS. The
problem is that it is not possible to open STDIN FD when using cat with HEREDOC. This should be
patched by rkapl so we are waiting for getting it to mainline.