Fix use after free when releasing memory.
We did not keep our promise and did not detach all virts when freeing a
phys. Also there could be an assertion (removing from list when not
added) in virt_do_free.

This also trggered another bug in decommit_virt, where conected_through
(the new, removed connection) was used instead of commited_to.