Netlink communication cleanup.
There was a lot of code duplication in nl.c which was reduced.
This is the first part of the code cleanup, the plan is to add
logging in the netlink layer (mostly for error reporting) and
to take into account the new return types (lsdn_err_t) in the
rest of the code base.