Test multiple settings in 1 network config file.
The possibility to specify multiple network
settings in 1 config file was basically already
done, added test for this use case. During testing
also revealed a problem when specifying e.g.:
settings A vxlan/e2e
settings B vxlan/static.
These specific settings lead to a crash of lsdn.

We should also improve the lsctl part so that
when we create vxlan/static settings A and
vxlan/static settings B then we want to detect
that these two settings are of the same type
and reject this input.