Prepare infrastructure for the firewall.
The main goal is to allow merging rules between the firewall and the TC
forwarding. So lsdn_ruleset was changed to take care of one whole chain,
with the old ruleset (flower filter) roughly corresponding to
lsdn_ruleset_prio. Each lsdn_ruleset_prio is extended so that it can
recognize multiple identical rules and merge their action lists (the
rules now have subpriority that governs the ordering inside the action
list).

This mechanism will be used for the firewall, the user will be given the
option to create rules in the lsdn_ruleset_prio that the virt uses for
routing and the lsdn_ruleset machinery will take care of joining them.

This also meant a change how a rule is defined. The lsdn_rule now must
contain all information needed for creating the rule. Previously it only
needed to track the information needed for deleting the rule.

To identify the identicall rules, we now need a hash table. I've chosen
uthash, it seems to be standard on both Arch and Buntu, sufficiently
lightweight and does no insist on keys being just strings.