Pass IP version to the kernel for vxlan device when needed.
When an IPv6 vxlan tunnel is created and neither group nor
remote is specified the kernel defaults to IPv4. So we need
to add the ipv explicitly. Similar fix can also be found in
iproute2 (commit 97d564b90cc).

Also had to increase the ping timeout, otherwise the large test
fails. It would be a good idea to write some tool which could be
added in the test framework and which would check whether there
isn't some extra traffic generated by us.

This commit (hopefully) completes the IPv6 support in LSDN.