Complete the doxygen split into full and public.
Both doxygens are referencable by breathe and you can choose which one
to use (using the :project: argument). The Doxyfiles are constructed
with a common doxyfile template.

Since we need both doxygens on ReadTheDocs, the conf.py now calls down
back to ./build-doc.sh to generate the Doxyfiles and run doxygen.