Lift the restriction of net and virt name being mandatory.
Previously, when the names were automatically generated
and you created some model with the config files, dumped the
model and then tried to reload the dumped model again it
happened that you could get a LSDNE_DUPLICATE error.

Currently the way dumping works net name and virt name
are not needed. We just need to have the phys name for the
case when you build a model with the C API and create a phys
without a name, but attach it to some network and then
lsdn_dump_context_tcl().