Basic geneve e2e (learning bridge) is working.
Got the broadcast problem solved by skbmarking packets and then
filtering them out by fw filter rule. Flower does not support skbmark.

Anyway: the problem was that the network topology with end-to-end
connecitons forms a complete graph. This leads to cycles which trip up
the bridges. So broadcast coming through a tunnel should not be
re-broadcast to other tunnels.